+++ package.json
@@ -8,7 +8,7 @@
     "start": "npm run build && node dist/main.js",
     "start:reuse": "npm run build && node dist/main.js --reuse-segs",
     "start:reuse:dir": "npm run build && node dist/main.js --reuse-segs --segsDir src/temp",
-    "test": "npm run build && node --test dist/**/*.test.js"
+    "test": "npm run build && node --test dist/tests"
   },
   "dependencies": {
     "ffmpeg-static": "^5.2.0"
